+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Print license (`pyframe.print_license()`)
 - Added [contributing guidelines](https://gitlab.com/FraME-projects/PyFraME/blob/master/CONTRIBUTING.MD)
 - Cost-effective polarizable protein potential (CP3) (see: https://doi.org/10.26434/chemrxiv.8126912)
+- The PDB input file reader will now try to guess the element based on the atom name, when the element field is missing.
 
 ### Fixed